Closes i-RIC/prepost-gui#370

+++ libs/tmsloader/tmsutil.h
@@ -12,7 +12,7 @@ namespace tmsloader {
 class TMSLOADER_API TmsUtil
 {
 public:
-	static double meterPerPixel(const QPointF &lonlat, int zoomLevel);
+	static double meterPerPixel(const QPointF &lonlat, double zoomLevel);
 	static int calcNativeZoomLevel(const QPointF& lonLat, double meterPerPixel);
 	/// Calculate the width and height of the image to cover the specified region and zoom level.
 	static QSize calcPixelSize(const QPointF& lonlat, const QSizeF& size, int zoomLevel);
